Error corrected for subitems in multi files
When comparing multiple files from different servers
and the server are just matching for one subitem, e.g.
IFACE lo then there was an error assuming that only
one sub device is not appropriate for metrics with multiple
devices. The change was
  if len(sub_items) > 1: set to if len(sub_items) >= 1:
else no subdevice could be elected and an error would be thrown.